refactor(ia-360): catalogo por dominio para times paralelos nao colidirem
Cinco waves entregando capacidades ao mesmo tempo editariam todas o mesmo array
de catalog.ts. Conflito em todo hunk, e conflito resolvido no automatico e onde
uma capacidade some em silencio.

Cada dominio passa a declarar no proprio arquivo em lib/mcp/tools/catalogo/;
o agregador so junta. Adicionar um dominio custa uma linha de import e uma de
spread — nenhum time toca o arquivo do outro.

catalog.ts continua existindo como porta de entrada: varios modulos ja importam
daquele caminho e o contrato publico nao muda.

Guarda nova contra o acidente provavel desse arranjo: dois dominios declarando
o mesmo `name`. Uma tool sombrearia a outra sem erro. Acusa em dev pelo
agregador e no CI pelo teste.
